requirements.txt
flloat==0.3.0
lark_parser==0.12.0
matplotlib==3.8.0
numpy==2.2.6
PyYAML==6.0.2
torch==2.5.1
torchvision==0.20.1
